Add hagezi Ultimate list
Not sure if there’s a way to move this right under pro++ and above the threat intelligence feeds list but but right there would be great. I can move id’s around so it goes there if that’s possible. Anyways, hagezi ultimate has all the pro++ and threat intelligence feeds lists but adds onto that so it’s more aggressive. Not sure how that’s possible with the amount it had already but here we are. If there’s anything I can, just let me know