More blocklist recommendations
USOM is made by the same person as Inversion but detects malware from Turkish feeds (or something like that). Duck duck go tracker radar is included with Blokada and this is one of the only lists that are not here that Blokada has. Nextdns recommended is 3 lists already included here combined but seems like a popular pick among nextdns users so it might be worth adding here anyways, RooneyMcNibNug’s SNAFU list is one that’s included on firebog and seems decent. 

These are the last lists i would recommend adding to dnswarden other than hagezi’s ultimate but we can wait a few weeks or months like you said👌.